unit2ixl-npu/requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.12.14
charset-normalizer==3.4.0
datasets==3.2.0
decorator==4.4.2
deepspeed==0.16.2
diffusers==0.31.0
dill==0.3.8
einops==0.8.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
hjson==3.1.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.3
numpy==1.24.0
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.0.0
propcache==0.2.1
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.10.1
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.3
tokenizers==0.20.4
torchvision==0.16.0
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
